+++ slash/__init__.py
@@ -34,7 +34,7 @@ from .assertions import (
 from .core.test import Test
 from .core.test import abstract_test_class
 from .core.fixtures import parametrize, parameters
-from .core.fixtures.utils import fixture, nofixtures, generator_fixture, yield_fixture
+from .core.fixtures.utils import fixture, nofixtures, generator_fixture, yield_fixture, use
 from .core.requirements import requires
 from .utils import skip_test, skipped, add_error, add_failure, set_test_detail, repeat
 from .utils.interactive import start_interactive_shell