+++ tests/utils/__init__.py
@@ -13,6 +13,7 @@ import pytest
 
 import gossip
 import slash
+from slash._compat import PY2, PYPY
 from slash.conf import config
 from slash._compat import ExitStack
 from slash.core.runnable_test import RunnableTest
@@ -139,3 +140,15 @@ def raises_maybe(exc, cond):
     if cond:
         return pytest.raises(exc)
     return noop()
+
+_noop = lambda f: f
+
+if PY2:
+    skip_on_py2 = pytest.mark.skip
+else:
+    skip_on_py2 = _noop
+
+if PYPY:
+    skip_on_pypy = pytest.mark.skip
+else:
+    skip_on_pypy = _noop